refactored flash app structure
